pyproject.toml
[project]
name = "alloy-shopify-slack"
version = "0.1.0"
description = "Shopify → Slack demo using Alloy's Connectivity API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
